Exercise 4.21

+++ part4/BlogList/controllers/blogs.js
@@ -42,8 +42,32 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
-    return response.sendStatus(204).end();
+    const blog = await Blog.findById(request.params.id);
+
+    if (!request.token) {
+        return response.status(401).json({
+            error: "token missing",
+        });
+    }
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+        return response.status(401).json({
+            error: "token invalid",
+        });
+    }
+
+    const user = await User.findById(decodedToken.id);
+
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id);
+        return response.sendStatus(204).end();
+    } else {
+        return response.status(401).json({
+            error: "you do not have permission to delete this blog",
+        });
+    }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
